+++ smart-pos-core/src/main/java/com/interswitchng/smartpos/shared/interfaces/device/EmvCardReader.kt
@@ -1,8 +1,6 @@
 package com.interswitchng.smartpos.shared.interfaces.device
 
-import com.interswitchng.smartpos.shared.interfaces.library.EmvCallback
 import com.interswitchng.smartpos.shared.models.core.TerminalInfo
-import com.interswitchng.smartpos.shared.models.transaction.cardpaycode.CardDetail
 import com.interswitchng.smartpos.shared.models.transaction.cardpaycode.EmvMessage
 import com.interswitchng.smartpos.shared.models.transaction.cardpaycode.EmvResult
 import com.interswitchng.smartpos.shared.models.transaction.cardpaycode.request.EmvData
@@ -30,7 +28,7 @@ interface EmvCardReader {
      * @see TerminalInfo
      * @see EmvMessage
      */
-    suspend fun setupTransaction(amount: Double, terminalInfo: TerminalInfo, channel: Channel<EmvMessage>, scope: CoroutineScope)
+    suspend fun setupTransaction(amount: Int, terminalInfo: TerminalInfo, channel: Channel<EmvMessage>, scope: CoroutineScope)
 
 
     /**
