+++ smart-pos-emv-pax/src/main/java/com/interswitchng/smartpos/emv/pax/emv/EmvImplementation.kt
@@ -22,6 +22,8 @@ import com.pax.jemv.emv.model.EmvMCKParam
 import com.pax.jemv.emv.model.EmvParam
 import kotlinx.coroutines.runBlocking
 import java.util.*
+import kotlin.ByteArray
+import kotlin.let
 
 
 /**
@@ -42,9 +44,9 @@ internal class EmvImplementation(private val context: Context, private val pinCa
     private lateinit var terminalInfo: TerminalInfo
 
 
-    private var amount: Double = 0.00
+    private var amount: Int = 0
 
-    fun setAmount(amount: Double) {
+    fun setAmount(amount: Int) {
         this.amount = amount
     }
